B2Ctinyshopapiauto/requirements.txt
pytest>=7.0.0
requests>=2.28.0
PyYAML>=6.0
jsonpath-ng>=1.5.3
pytest-html>=3.1.0
pytest-xdist>=3.0.0
allure-python-commons~=2.13.2